Append generated tests to test file if it already exists, and import headers from file-under-test.

+++ output/output.go
@@ -11,7 +11,7 @@ import (
 	"golang.org/x/tools/imports"
 )
 
-func Write(srcPath, destPath string, head *models.Header, funcs []*models.Function) ([]string, error) {
+func Write(srcPath, destPath string, head models.Header, funcs []*models.Function) ([]string, error) {
 	tf, err := ioutil.TempFile("", "gotests_")
 	if err != nil {
 		return nil, fmt.Errorf("ioutil.TempFile: %v", err)
@@ -21,19 +21,32 @@ func Write(srcPath, destPath string, head *models.Header, funcs []*models.Functi
 	if err != nil {
 		return nil, err
 	}
-	df, err := os.Create(destPath)
-	if err != nil {
-		return nil, fmt.Errorf("os.Create: %v", err)
+	var isDestNew bool
+	if IsFileExist(destPath) {
+		df, err := os.Create(destPath)
+		if err != nil {
+			return nil, fmt.Errorf("os.Create: %v", err)
+		}
+		defer df.Close()
+		// df.Chmod(os.ModePerm)
+		destPath = df.Name()
+		isDestNew = true
 	}
-	defer df.Close()
-	if err := copyTempToDest(tf, df); err != nil {
-		os.Remove(df.Name())
+	if err := copyTempToDest(tf.Name(), destPath); err != nil {
+		if isDestNew {
+			os.Remove(destPath)
+		}
 		return nil, err
 	}
 	return tests, nil
 }
 
-func writeTestsToTemp(temp *os.File, head *models.Header, funcs []*models.Function) ([]string, error) {
+func IsFileExist(path string) bool {
+	_, err := os.Stat(path)
+	return !os.IsNotExist(err)
+}
+
+func writeTestsToTemp(temp *os.File, head models.Header, funcs []*models.Function) ([]string, error) {
 	w := bufio.NewWriter(temp)
 	if err := render.Header(w, head); err != nil {
 		return nil, fmt.Errorf("render.Header: %v", err)
@@ -43,44 +56,33 @@ func writeTestsToTemp(temp *os.File, head *models.Header, funcs []*models.Functi
 		if err := render.TestFunction(w, fun); err != nil {
 			return nil, fmt.Errorf("render.TestFunction: %v", err)
 		}
-		tests = append(tests, fmt.Sprintf("%v.%v", head.Package, fun.TestName()))
+		tests = append(tests, fmt.Sprintf("%v.%v", head.Package(), fun.TestName()))
 	}
 	if err := w.Flush(); err != nil {
 		return nil, fmt.Errorf("bufio.Flush: %v", err)
 	}
-	if err := processImports(temp); err != nil {
+	if err := processImports(temp.Name()); err != nil {
 		return nil, fmt.Errorf("processImports: %v", err)
 	}
 	return tests, nil
 }
 
-func processImports(f *os.File) error {
-	v, err := ioutil.ReadFile(f.Name())
+func processImports(path string) error {
+	v, err := ioutil.ReadFile(path)
 	if err != nil {
 		return fmt.Errorf("ioutil.ReadFile: %v", err)
 	}
-	b, err := imports.Process(f.Name(), v, nil)
+	b, err := imports.Process(path, v, nil)
 	if err != nil {
 		return fmt.Errorf("imports.Process: %v", err)
 	}
-	return overwriteFile(f, b)
+	return ioutil.WriteFile(path, b, 0)
 }
 
-func copyTempToDest(tempf, destf *os.File) error {
-	b, err := ioutil.ReadFile(tempf.Name())
+func copyTempToDest(tempPath, destPath string) error {
+	b, err := ioutil.ReadFile(tempPath)
 	if err != nil {
 		return fmt.Errorf("ioutil.ReadAll: %v", err)
 	}
-	return overwriteFile(destf, b)
-}
-
-func overwriteFile(f *os.File, b []byte) error {
-	n, err := f.WriteAt(b, 0)
-	if err != nil {
-		return fmt.Errorf("file.Write: %v", err)
-	}
-	if err := f.Truncate(int64(n)); err != nil {
-		return fmt.Errorf("file.Truncate: %v", err)
-	}
-	return f.Sync()
+	return ioutil.WriteFile(destPath, b, os.ModePerm)
 }
